requirements.txt
Flask
Flask-SocketIO
gunicorn
opencv-python
eventlet
pymongo
mongoengine
graphene
graphene-mongo
Flask-GraphQL
python-dotenv